# Use SipHash-1-3 for all hashing, through a resilient hashing interface