--- a/stdlib/public/core/StringHashable.swift
+++ b/stdlib/public/core/StringHashable.swift
@@ -100,7 +100,7 @@ extension _UnmanagedString where CodeUnit == UInt8 {
 #else
     var hasher = _Hasher()
     Unicode.hashASCII(self.buffer, into: &hasher)
-    return hasher.finalize()
+    return Int(truncatingIfNeeded: hasher.finalize())
 #endif // _runtime(_ObjC)
   }
 }
@@ -120,7 +120,7 @@ extension _UnmanagedString where CodeUnit == UTF16.CodeUnit {
 #else
     var hasher = _Hasher()
     Unicode.hashUTF16(self.buffer, into: &hasher)
-    return hasher.finalize()
+    return Int(truncatingIfNeeded: hasher.finalize())
 #endif // _runtime(_ObjC)
   }
 }
@@ -145,7 +145,7 @@ extension _UnmanagedOpaqueString {
     Unicode.hashUTF16(
       UnsafeBufferPointer(start: p, count: count),
       into: &hasher)
-    return hasher.finalize()
+    return Int(truncatingIfNeeded: hasher.finalize())
 #endif
   }
 }