… stdlib/public/core/SipHash.swift.gyb
… test/SILGen/objc_bridging_any.swift
… validation-test/stdlib/Hashing.swift
… validation-test/stdlib/HashingAvalanche.swift
… validation-test/stdlib/SipHash.swift