@lorentey 2018-03-01
@swift-ci please test