@milseman 2018-03-01
@atrick @aschwaighofer is this ok usage of `@effects(readonly)`?