@jrose-apple 2018-03-01
> is this ok usage of `@effects(readonly)`?

I'm not Andy or Arnold but I say definitely not. `@effects(readonly)` means that the optimizer can omit the call to the function if its return value isn't used. Even if we don't do that optimization today I really wouldn't want to put it in that way. We can invent something else to avoid retain/releasing the Hasher, or just cross our fingers for MichaelG's +0 work.