@jrose-apple 2018-03-01
Question: why isn't `_UnsafeHasher` taken as `inout`, rather than taken/returned?