@lorentey 2018-03-01
The return value is actually being used here, though; the calls are chained together to prevent invalid optimizations. The hasher state itself is just a bunch of integers; `@effects(readonly)` prevents ARC traffic protecting components of the `Hashable` types, not the hasher itself.