@aschwaighofer 2018-03-01
@milseman (@atrick, @eeckstein ) Like you suspected -- I don't think we would want to annotate a function that mutates state as 'readonly' even if it happens that within our existing pipeline this might work for you.

I think we need a ``@effects(releasenone)`` and map that to sideeffectsanalysis' 'GlobalEffects.Release=false'. I think that should trick. @eeckstein?