@lorentey 2018-03-01
`inout` implies mutations, and every time I tried providing **opaque** mutating functions on `Hasher`, I measured major ARC-related slowdowns (as much as +300%, if I recall correctly). Inlineable mutating funcs were fine, but as soon as the body wasn't available, things got really slow.

I am yet to try running this on the +0 branch, but it sure would be nice to replace this with a straightforward interface.