@lorentey 2018-03-01
@aschwaighofer Ah, `GlobalEffects.mayRelease()` explains so much; it seems like a new `@effects` attribute for it would fit the bill perfectly here, and we'd probably use it elsewhere throughout the stdlib.

Replacing the fragile `_UnsafeHasher` hack with an `inout Hasher` would allow us to consider promoting `hash(into:)` into a public replacement for `hashValue`.