@aschwaighofer 2018-03-01
@lorenty I have prototyped this in the following branch:
https://github.com/aschwaighofer/swift/tree/wip_effects_release_none