@eeckstein 2018-03-01
> I think we need a @effects(releasenone) and map that to sideeffectsanalysis' 'GlobalEffects.Release=false'. I think that should trick. @eeckstein?

Yes, that should work.