@lorentey 2018-03-02
While we have the tests passing, let's see if anyone relies on specific hash values:

@swift-ci please test source compatibility