@lorentey 2018-03-04
Okay, my own benchmark results for bd782d79a7e44a5292104ab86392fdd286614c5f in #14968 got the following results. It's not too bad, but it's still a bit slower than `_UnsafeHasher`. I'll do a detailed analysis soon.

<details open>
  <summary>Regression (75)</summary>
  
TEST                                                      | OLD    | NEW    | DELTA   | SPEEDUP 
---                                                       | ---    | ---    | ---     | ---     
SetIntersect                                              | 275    | 911    | +231.3% | **0.30x** 
DictionarySwap                                            | 410    | 1187   | +189.5% | **0.35x** 
ReversedDictionary                                        | 124    | 328    | +164.5% | **0.38x** 
DictionaryGroup                                           | 109    | 270    | +147.7% | **0.40x** 
Histogram                                                 | 267    | 644    | +141.2% | **0.41x** 
Dictionary4                                               | 127    | 303    | +138.6% | **0.42x** 
Dictionary4OfObjects                                      | 187    | 427    | +128.3% | **0.44x** 
DictionarySubscriptDefaultMutation                        | 133    | 286    | +115.0% | **0.47x** 
SuffixCountableRange                                      | 6      | 12     | +100.0% | **0.50x** 
SuffixCountableRangeLazy                                  | 6      | 11     | +83.3%  | **0.55x** 
SetExclusiveOr                                            | 3115   | 5541   | +77.9%  | **0.56x** 
DictionaryRemove                                          | 2975   | 5107   | +71.7%  | **0.58x** 
SetIntersect_OfObjects                                    | 1759   | 2932   | +66.7%  | **0.60x** 
SetUnion                                                  | 2824   | 4704   | +66.6%  | **0.60x** 
CharIteration_utf16_unicodeScalars_Backwards              | 14363  | 23687  | +64.9%  | **0.61x** 
RGBHistogram                                              | 1976   | 3121   | +57.9%  | **0.63x** 
TwoSum                                                    | 928    | 1463   | +57.7%  | **0.63x** 
SetExclusiveOr_OfObjects                                  | 8593   | 13085  | +52.3%  | **0.66x** 
Prims                                                     | 723    | 1069   | +47.9%  | **0.68x** 
StackPromo                                                | 22261  | 32868  | +47.6%  | **0.68x** 
PrimsSplit                                                | 726    | 1054   | +45.2%  | **0.69x** 
SortStringsUnicode                                        | 2207   | 3204   | +45.2%  | **0.69x** 
SetUnion_OfObjects                                        | 7003   | 10040  | +43.4%  | **0.70x** 
CharacterPropertiesPrecomputed                            | 934    | 1299   | +39.1%  | **0.72x** 
CharacterLiteralsLarge                                    | 5911   | 8040   | +36.0%  | **0.74x** 
WordCountUniqueASCII                                      | 1630   | 2159   | +32.5%  | **0.75x** 
ObjectiveCBridgeStubToNSStringRef                         | 88     | 116    | +31.8%  | **0.76x** 
Dictionary3                                               | 221    | 289    | +30.8%  | **0.76x** 
ClassArrayGetter                                          | 10     | 13     | +30.0%  | **0.77x** 
SetIsSubsetOf_OfObjects                                   | 396    | 510    | +28.8%  | **0.78x** 
EqualStringSubstring                                      | 36     | 46     | +27.8%  | **0.78x** 
PrefixCountableRange                                      | 18     | 23     | +27.8%  | **0.78x** 
RangeIterationSigned                                      | 115    | 145    | +26.1%  | **0.79x** 
StaticArray                                               | 8      | 10     | +25.0%  | **0.80x** 
DictionarySubscriptDefaultMutationArray                   | 497    | 621    | +24.9%  | **0.80x** 
Dictionary                                                | 461    | 571    | +23.9%  | **0.81x** 
SubstringComparable                                       | 51     | 63     | +23.5%  | **0.81x** 
DictionaryLiteral                                         | 1595   | 1952   | +22.4%  | **0.82x** 
CharacterPropertiesStashedMemo                            | 1438   | 1748   | +21.6%  | **0.82x** 
DictionaryGroupOfObjects                                  | 1920   | 2294   | +19.5%  | **0.84x** 
ObjectiveCBridgeStubFromNSStringRef                       | 143    | 170    | +18.9%  | **0.84x** 
DictionarySwapOfObjects                                   | 7832   | 9201   | +17.5%  | **0.85x** 
ArrayOfPOD                                                | 122    | 143    | +17.2%  | **0.85x** 
PointerArithmetics                                        | 23996  | 27827  | +16.0%  | **0.86x** 
DictionaryRemoveOfObjects                                 | 24497  | 28328  | +15.6%  | **0.86x** 
SetIsSubsetOf                                             | 316    | 365    | +15.5%  | **0.87x** 
DictionarySubscriptDefaultMutationOfObjects               | 1466   | 1671   | +14.0%  | **0.88x** 
CharIteration_utf16_unicodeScalars                        | 25833  | 29292  | +13.4%  | **0.88x** 
RGBHistogramOfObjects                                     | 24003  | 27177  | +13.2%  | **0.88x** 
SuffixAnySeqCntRange                                      | 24     | 27     | +12.5%  | **0.89x** 
StringAdder                                               | 4318   | 4789   | +10.9%  | **0.90x** 
StrComplexWalk                                            | 1356   | 1499   | +10.5%  | **0.90x** 
StringComparison_abnormal                                 | 673    | 741    | +10.1%  | **0.91x** 
DictionaryBridge                                          | 1519   | 1667   | +9.7%   | **0.91x** 
ObjectiveCBridgeStubFromNSString                          | 710    | 777    | +9.4%   | **0.91x** 
Dictionary2                                               | 1523   | 1661   | +9.1%   | **0.92x** 
EqualSubstringString                                      | 36     | 39     | +8.3%   | **0.92x** 
NSStringConversion                                        | 246    | 266    | +8.1%   | **0.92x** 
CharIteration_korean_unicodeScalars                       | 16755  | 18053  | +7.7%   | **0.93x** 
StrToInt                                                  | 1709   | 1840   | +7.7%   | **0.93x** 
ArrayOfGenericPOD2                                        | 125    | 134    | +7.2%   | **0.93x** 
Dictionary2OfObjects                                      | 2864   | 3062   | +6.9%   | **0.94x** 
PrefixArray                                               | 30     | 32     | +6.7%   | **0.94x** 
RangeAssignment                                           | 266    | 283    | +6.4%   | **0.94x** 
ObjectiveCBridgeStubToNSDate                              | 15488  | 16454  | +6.2%   | **0.94x** 
OpenClose                                                 | 97     | 103    | +6.2%   | **0.94x** 
StringWordBuilder                                         | 1579   | 1675   | +6.1%   | **0.94x** 
CharIteration_russian_unicodeScalars                      | 14035  | 14882  | +6.0%   | **0.94x** 
FrequenciesUsingReduceInto                                | 3106   | 3293   | +6.0%   | **0.94x** 
SortStrings                                               | 1544   | 1635   | +5.9%   | **0.94x** 
ReversedArray                                             | 51     | 54     | +5.9%   | **0.94x** 
CharIteration_chinese_unicodeScalars                      | 12729  | 13461  | +5.8%   | **0.95x** 
LessSubstringSubstring                                    | 35     | 37     | +5.7%   | **0.95x** 
ArrayAppendUTF16                                          | 29557  | 31228  | +5.7%   | **0.95x** 
Array2D                                                   | 1924   | 2030   | +5.5%   | **0.95x** 

</details>
<details >
  <summary>Improvement (33)</summary>
  
TEST                                                      | OLD    | NEW    | DELTA   | SPEEDUP 
---                                                       | ---    | ---    | ---     | ---     
PrefixCountableRangeLazy                                  | 36     | 18     | -50.0%  | **2.00x** 
PrefixWhileCountableRangeLazy                             | 35     | 18     | -48.6%  | **1.94x** 
WordSplitUTF16                                            | 20183  | 12139  | -39.9%  | **1.66x** 
WordSplitASCII                                            | 18938  | 13520  | -28.6%  | **1.40x** 
PrefixWhileCountableRange                                 | 36     | 28     | -22.2%  | **1.29x** 
RangeIterationSigned64                                    | 146    | 116    | -20.5%  | **1.26x** 
RangeIterationUnsigned                                    | 146    | 116    | -20.5%  | **1.26x** 
DropFirstArrayLazy                                        | 37     | 30     | -18.9%  | **1.23x** 
Chars                                                     | 766    | 642    | -16.2%  | **1.19x** 
ArrayAppendLazyMap                                        | 959    | 827    | -13.8%  | **1.16x** 
BitCount                                                  | 207    | 184    | -11.1%  | **1.12x** 
ArrayAppendAscii                                          | 11475  | 10354  | -9.8%   | **1.11x** 
ObjectiveCBridgeFromNSSetAnyObject                        | 54558  | 49863  | -8.6%   | **1.09x** 
ObjectiveCBridgeFromNSString                              | 1381   | 1267   | -8.3%   | **1.09x** 
StringEnumRawValueInitialization                          | 948    | 870    | -8.2%   | **1.09x** 
AnyHashableWithAClass                                     | 73599  | 68154  | -7.4%   | **1.08x** 
DropLastSequence                                          | 496    | 461    | -7.1%   | **1.08x** 
DropWhileSequenceLazy                                     | 48     | 45     | -6.2%   | **1.07x** 
RemoveWhereSwapInts                                       | 16     | 15     | -6.2%   | **1.07x** 
StringWordBuilderReservingCapacity                        | 1087   | 1020   | -6.2%   | **1.07x** 
LazilyFilteredArrayContains                               | 32184  | 30253  | -6.0%   | **1.06x** 
PrefixArrayLazy                                           | 36     | 34     | -5.6%   | **1.06x** 
StringHasSuffixAscii                                      | 1044   | 988    | -5.4%   | **1.06x** 
ArraySubscript                                            | 1556   | 1473   | -5.3%   | **1.06x** 
DropWhileCountableRange                                   | 19     | 18     | -5.3%   | **1.06x** 
CharIndexing_punctuated_unicodeScalars                    | 2845   | 2701   | -5.1%   | **1.05x** 
SequenceAlgosArray                                        | 1759   | 1670   | -5.1%   | **1.05x** 
CharIndexing_russian_unicodeScalars_Backwards             | 9694   | 9205   | -5.0%   | **1.05x** 
CharIndexing_japanese_unicodeScalars_Backwards            | 13925  | 13228  | -5.0%   | **1.05x** 
CharIndexing_russian_unicodeScalars                       | 10358  | 9849   | -4.9%   | **1.05x** 
CharIndexing_japanese_unicodeScalars                      | 14871  | 14141  | -4.9%   | **1.05x** 
NopDeinit                                                 | 21566  | 20519  | -4.9%   | **1.05x** 
StringComparison_nonBMPSlowestPrenormal                   | 1364   | 1298   | -4.8%   | **1.05x** 

</details>

<details >
  <summary>No Changes (270)</summary>
  
TEST                                                      | OLD    | NEW    | DELTA   | SPEEDUP 
---                                                       | ---    | ---    | ---     | ---     
AngryPhonebook                                            | 3570   | 3582   | +0.3%   | 1.00x   
ArrayAppend                                               | 981    | 980    | -0.1%   | 1.00x   
ArrayAppendArrayOfInt                                     | 424    | 441    | +4.0%   | 0.96x   
ArrayAppendFromGeneric                                    | 427    | 440    | +3.0%   | 0.97x   
ArrayAppendGenericStructs                                 | 843    | 851    | +0.9%   | 0.99x   
ArrayAppendLatin1                                         | 30295  | 31528  | +4.1%   | 0.96x   
ArrayAppendOptionals                                      | 837    | 860    | +2.7%   | 0.97x   
ArrayAppendRepeatCol                                      | 848    | 860    | +1.4%   | 0.99x   
ArrayAppendReserved                                       | 806    | 794    | -1.5%   | 1.02x   
ArrayAppendSequence                                       | 746    | 753    | +0.9%   | 0.99x   
ArrayAppendStrings                                        | 14226  | 14259  | +0.2%   | 1.00x   
ArrayAppendToFromGeneric                                  | 418    | 429    | +2.6%   | 0.97x   
ArrayAppendToGeneric                                      | 434    | 427    | -1.6%   | 1.02x   
ArrayInClass                                              | 61     | 62     | +1.6%   | 0.98x   
ArrayLiteral                                              | 0      | 0      | +0.0%   | 1.00x   
ArrayOfGenericRef                                         | 4467   | 4592   | +2.8%   | 0.97x   
ArrayOfRef                                                | 4413   | 4508   | +2.2%   | 0.98x   
ArrayPlusEqualArrayOfInt                                  | 435    | 439    | +0.9%   | 0.99x   
ArrayPlusEqualFiveElementCollection                       | 5184   | 5261   | +1.5%   | 0.99x   
ArrayPlusEqualSingleElementCollection                     | 989    | 1002   | +1.3%   | 0.99x   
ArrayPlusEqualThreeElements                               | 1652   | 1700   | +2.9%   | 0.97x   
ArrayValueProp                                            | 5      | 5      | +0.0%   | 1.00x   
ArrayValueProp2                                           | 5      | 5      | +0.0%   | 1.00x   
ArrayValueProp3                                           | 5      | 5      | +0.0%   | 1.00x   
ArrayValueProp4                                           | 5      | 5      | +0.0%   | 1.00x   
BinaryFloatingPointConversionFromBinaryInteger            | 28     | 28     | +0.0%   | 1.00x   
ByteSwap                                                  | 87     | 86     | -1.1%   | 1.01x   
COWTree                                                   | 5452   | 5486   | +0.6%   | 0.99x   
CSVParsing                                                | 752371 | 743742 | -1.1%   | 1.01x   
CSVParsingAlt                                             | 687421 | 700616 | +1.9%   | 0.98x   
CSVParsingAltIndices                                      | 367034 | 360282 | -1.8%   | 1.02x   
CStringLongAscii                                          | 3416   | 3383   | -1.0%   | 1.01x   
CStringLongNonAscii                                       | 1595   | 1637   | +2.6%   | 0.97x   
CStringShortAscii                                         | 4578   | 4464   | -2.5%   | 1.03x   
Calculator                                                | 490    | 481    | -1.8%   | 1.02x   
CaptureProp                                               | 2717   | 2722   | +0.2%   | 1.00x   
CharIndexing_ascii_unicodeScalars                         | 12355  | 11792  | -4.6%   | 1.05x   
CharIndexing_ascii_unicodeScalars_Backwards               | 11661  | 11162  | -4.3%   | 1.04x   
CharIndexing_chinese_unicodeScalars                       | 9389   | 8969   | -4.5%   | 1.05x   
CharIndexing_chinese_unicodeScalars_Backwards             | 8820   | 8515   | -3.5%   | 1.04x   
CharIndexing_korean_unicodeScalars                        | 12003  | 11546  | -3.8%   | 1.04x   
CharIndexing_korean_unicodeScalars_Backwards              | 11260  | 10727  | -4.7%   | 1.05x   
CharIndexing_punctuatedJapanese_unicodeScalars            | 2269   | 2207   | -2.7%   | 1.03x   
CharIndexing_punctuatedJapanese_unicodeScalars_Backwards  | 2137   | 2086   | -2.4%   | 1.02x   
CharIndexing_punctuated_unicodeScalars_Backwards          | 2684   | 2589   | -3.5%   | 1.04x   
CharIndexing_tweet_unicodeScalars                         | 24466  | 24961  | +2.0%   | 0.98x   
CharIndexing_tweet_unicodeScalars_Backwards               | 22836  | 21882  | -4.2%   | 1.04x   
CharIndexing_utf16_unicodeScalars                         | 19291  | 18712  | -3.0%   | 1.03x   
CharIndexing_utf16_unicodeScalars_Backwards               | 19731  | 19662  | -0.3%   | 1.00x   
CharIteration_ascii_unicodeScalars                        | 17041  | 17293  | +1.5%   | 0.99x   
CharIteration_ascii_unicodeScalars_Backwards              | 10441  | 10428  | -0.1%   | 1.00x   
CharIteration_chinese_unicodeScalars_Backwards            | 8115   | 8018   | -1.2%   | 1.01x   
CharIteration_japanese_unicodeScalars                     | 20795  | 20927  | +0.6%   | 0.99x   
CharIteration_japanese_unicodeScalars_Backwards           | 12597  | 12803  | +1.6%   | 0.98x   
CharIteration_korean_unicodeScalars_Backwards             | 10232  | 10663  | +4.2%   | 0.96x   
CharIteration_punctuatedJapanese_unicodeScalars           | 3089   | 3174   | +2.8%   | 0.97x   
CharIteration_punctuatedJapanese_unicodeScalars_Backwards | 1942   | 1944   | +0.1%   | 1.00x   
CharIteration_punctuated_unicodeScalars                   | 3810   | 3887   | +2.0%   | 0.98x   
CharIteration_punctuated_unicodeScalars_Backwards         | 2387   | 2429   | +1.8%   | 0.98x   
CharIteration_russian_unicodeScalars_Backwards            | 8693   | 9041   | +4.0%   | 0.96x   
CharIteration_tweet_unicodeScalars                        | 33363  | 34353  | +3.0%   | 0.97x   
CharIteration_tweet_unicodeScalars_Backwards              | 20698  | 21397  | +3.4%   | 0.97x   
CharacterLiteralsSmall                                    | 338    | 348    | +3.0%   | 0.97x   
CharacterPropertiesFetch                                  | 4625   | 4635   | +0.2%   | 1.00x   
CharacterPropertiesStashed                                | 1500   | 1488   | -0.8%   | 1.01x   
Combos                                                    | 471    | 460    | -2.3%   | 1.02x   
DictOfArraysToArrayOfDicts                                | 773    | 791    | +2.3%   | 0.98x   
Dictionary3OfObjects                                      | 689    | 723    | +4.9%   | 0.95x   
DictionaryOfObjects                                       | 2172   | 2281   | +5.0%   | 0.95x   
DictionarySubscriptDefaultMutationArrayOfObjects          | 4002   | 4162   | +4.0%   | 0.96x   
DoubleWidthDivision                                       | 0      | 0      | +0.0%   | 1.00x   
DropFirstAnyCollection                                    | 69     | 70     | +1.4%   | 0.99x   
DropFirstAnyCollectionLazy                                | 71347  | 70681  | -0.9%   | 1.01x   
DropFirstAnySeqCRangeIter                                 | 18417  | 19137  | +3.9%   | 0.96x   
DropFirstAnySeqCRangeIterLazy                             | 18762  | 19049  | +1.5%   | 0.98x   
DropFirstAnySeqCntRange                                   | 62     | 61     | -1.6%   | 1.02x   
DropFirstAnySeqCntRangeLazy                               | 62     | 61     | -1.6%   | 1.02x   
DropFirstAnySequence                                      | 4410   | 4401   | -0.2%   | 1.00x   
DropFirstAnySequenceLazy                                  | 4434   | 4475   | +0.9%   | 0.99x   
DropFirstArray                                            | 31     | 30     | -3.2%   | 1.03x   
DropFirstCountableRange                                   | 18     | 18     | +0.0%   | 1.00x   
DropFirstCountableRangeLazy                               | 18     | 18     | +0.0%   | 1.00x   
DropFirstSequence                                         | 2952   | 2921   | -1.1%   | 1.01x   
…